% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.R
\name{methplot}
\alias{methplot}
\title{methplot}
\usage{
methplot(genome, chr, start, end, sites, regions, enhancers, group)
}
\arguments{
\item{genome}{Available reference genomes: hg19, hg38 and mm39.}

\item{chr}{Chromosome.}

\item{start}{First position (included).}

\item{end}{Last position (included).}

\item{sites}{GRanges object containing the CpG sites and its methylation values associated.}

\item{regions}{GRanges object or data.frame containing the detected DMRs.}

\item{enhancers}{data.frame containing the coordinates of enhancers (optional).}

\item{group}{Feature to group the samples (optional).}
}
\description{
Represents the methylation levels and the detected DMRs present in a defined region of the genome for a set of samples.
}
\examples{
genome <- "hg19"
chr <- "chr7"
start <- 14111939
end <- 15071940
group <- c("control", "cond_A", "cond_A", "cond_B", "control", "cond_A", ...)
}
\author{
Raul Sanz, \email{raulsanzr@gmail.com}
}
